+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.5.0] - 2017-10-18
+### Added
+- Allow create instances of not registered types: enable with `Injector::allowInstantiateNotRegisteredTypes(true)`.
+- Register Injector it self (useful for injection to factories): use `Injector::registerItself()`. 
+
 ## [1.4.0] - 2017-08-17
 ### Added
 - Allow to use default value if required type not set in container: enable with `Injector::allowNullableConstructorArgs(true)`.